Fix condition for uploading to Anaconda.org
Input is bool so no need to compare to anything. Previously, "If the types do not match, GitHub coerces the type to a number." therefore `inputs.upload_to_anaconda` which is bool `true` gets converted to `1`, and `'true'` which is a non-empty non-number string gets converted to `nan`. So it's `false` as `1` is not equal to `nan`.